Add a Collidable interface to the collision system
